fix: move Store to top of home filter panel; scope filters.css away from stack variant
