+++ app.html
@@ -15,7 +15,7 @@
 <link rel="stylesheet" href="css/shared/base.css?v=45c7e06d">
 <link rel="stylesheet" href="css/shared/topbar.css?v=55e85c86">
 <link rel="stylesheet" href="css/shared/site.css?v=4d518a51">
-<link rel="stylesheet" href="css/shared/filters.css?v=3b0c60ff">
+<link rel="stylesheet" href="css/shared/filters.css?v=4e3a1985">
 <link rel="stylesheet" href="css/shared/cards.css?v=68f59a72">
 <link rel="stylesheet" href="css/app/game-header.css?v=cec66a65">
 <link rel="stylesheet" href="css/app/configs.css?v=90d5bed5">
@@ -58,6 +58,6 @@
 <script src="js/lib/topbar.js?v=50b75e70"></script>
 <script src="js/lib/toast.js?v=7bb23014"></script>
 <script src="js/lib/gh-gist.js?v=346e3ea7"></script>
-<script type="module" src="js/app/main.js?v=96c048a9"></script>
+<script type="module" src="js/app/main.js?v=6f1d79d5"></script>
 </body>
 </html>
