+++ css/shared/filters.css
@@ -63,7 +63,7 @@
 /* ── Panel container ──────────────────────────────────────────────────── */
 /* app.html and stats.html use .filter-panel; index uses .pg-filter-panel.
    Both are absolutely-positioned dropdowns anchored to their toggle button. */
-.filter-panel,
+.filter-panel:not(.filter-panel--stack),
 .pg-filter-panel {
   position: absolute;
   top: calc(100% + 6px);
@@ -80,7 +80,7 @@
 
 /* ── Responsive ───────────────────────────────────────────────────────── */
 @media (min-width: 601px) {
-  .filter-panel,
+  .filter-panel:not(.filter-panel--stack),
   .pg-filter-panel {
     min-width: 480px;
   }
