+++ index.html
@@ -15,7 +15,7 @@
 <link rel="stylesheet" href="css/shared/base.css?v=45c7e06d">
 <link rel="stylesheet" href="css/shared/topbar.css?v=55e85c86">
 <link rel="stylesheet" href="css/shared/site.css?v=4d518a51">
-<link rel="stylesheet" href="css/shared/filters.css?v=3b0c60ff">
+<link rel="stylesheet" href="css/shared/filters.css?v=4e3a1985">
 <link rel="stylesheet" href="css/shared/cards.css?v=68f59a72">
 <link rel="stylesheet" href="css/index/index.css?v=ca60e7d4">
 </head>
